+++ test-scripts/qemu.py
@@ -3,12 +3,24 @@
 # See License.txt for details
 import os
 import sys
+import time
+from typing import List, Literal
 
 import pexpect
 
+SystemdState = Literal[
+    "inactive",
+    "active",
+    "deactivating",
+    "activating",
+    "reloading",
+    "failed",
+    "maintenance",
+]
+
 
 class QEMU:
-    def __init__(self):
+    def __init__(self, image_wic_path: os.PathLike):
         bitbake_path = os.getenv("BUILDDIR")
         assert bitbake_path, "Missing BUILDDIR environment variable"
 
@@ -27,9 +39,7 @@ class QEMU:
             "-object rng-random,filename=/dev/urandom,id=rng0 -device virtio-rng-pci,rng=rng0"
         )
         cmd_line_parts.append(
-            "-drive id=disk0,file="
-            + build_output_path
-            + "/base-image-qemuarm64.wic,if=none,format=raw -device virtio-blk-device,drive=disk0"
+            f"-drive id=disk0,file={image_wic_path},if=none,format=raw -device virtio-blk-device,drive=disk0"
         )
         cmd_line_parts.append(
             "-device qemu-xhci -device usb-tablet -device usb-kbd -device virtio-gpu-pci -nographic"
@@ -48,3 +58,37 @@ class QEMU:
 
     def child(self):
         return self.pid
+
+    def exec_cmd(self, cmd):
+        self.pid.expect("#")
+        self.pid.sendline(cmd)
+        self.pid.expect("\n")
+
+    def systemd_wait_for_service_state(
+        self,
+        service: str,
+        expected_state: SystemdState,
+        *,
+        timeout_seconds: float = 10.0,
+    ):
+        states: List[SystemdState] = [
+            "inactive",
+            "active",
+            "deactivating",
+            "activating",
+            "reloading",
+            "failed",
+            "maintenance",
+        ]
+        expected_state_idx = states.index(expected_state)
+        timeout = time.time() + timeout_seconds
+        last_state_idx = -1
+        while time.time() < timeout:
+            self.exec_cmd(f"systemctl is-active {service}")
+            last_state_idx = self.child().expect(states)
+            if last_state_idx == expected_state_idx:
+                return
+            time.sleep(0.1)
+        raise TimeoutError(
+            f"Timed out waiting for service {service} to get state {expected_state}. Last state: {states[last_state_idx]}"
+        )